feat: Complete project overhaul, security hardening, and Dockerization
This commit delivers a comprehensive overhaul of the full-stack e-commerce application, addressing key areas of security, code structure, and deployment.

Backend Refactoring:
- Restructured the monolithic Flask application into a scalable, modular architecture using the Application Factory pattern and Blueprints.
- Externalized configuration into a `config.py` file with support for different environments (development, production).
- Implemented `Flask-Migrate` for robust database schema management.

Security Hardening:
- Patched a critical client-side authorization vulnerability by removing reliance on `localStorage` for user authentication state. The frontend now validates the session with the server on load.
- Implemented full CSRF protection on the backend and integrated it with the frontend for all state-changing requests.
- Secured file uploads by validating file extensions and generating unique, safe filenames.

Deployment:
- Fully containerized the frontend and backend services using Docker.
- Created a `docker-compose.yml` to orchestrate the entire application stack (backend, frontend).
- Provided a production-ready Nginx configuration for serving the React frontend.
- Added comprehensive documentation to the `README.md` explaining the new structure and how to run the project using Docker Compose.
