+++ backend/app/api/routes.py
@@ -1,0 +1,14 @@
+from flask import Blueprint, jsonify
+from ..models import Product
+
+api_bp = Blueprint('api_bp', __name__)
+
+@api_bp.route('/products', methods=['GET'])
+def get_products():
+    products = Product.query.all()
+    return jsonify([product.to_dict() for product in products])
+
+@api_bp.route('/products/<int:product_id>', methods=['GET'])
+def get_product(product_id):
+    product = Product.query.get_or_404(product_id)
+    return jsonify(product.to_dict())
