+++ backend/app/models.py
@@ -1,0 +1,80 @@
+from .extensions import db
+from flask import url_for
+
+class Product(db.Model):
+    __tablename__ = 'products'
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(100), nullable=False)
+    price = db.Column(db.Float, nullable=False)
+    stock = db.Column(db.Integer, default=0)
+    description = db.Column(db.Text, nullable=True)
+    brand = db.Column(db.String(100), nullable=True)
+    images = db.relationship('ProductImage', backref='product', lazy=True, cascade="all, delete-orphan")
+
+    def to_dict(self):
+        image_urls = [url_for('static', filename=f'uploads/products/{image.filename}', _external=True) for image in self.images]
+        return {
+            'id': self.id,
+            'name': self.name,
+            'price': self.price,
+            'stock': self.stock,
+            'description': self.description,
+            'brand': self.brand,
+            'imageUrls': image_urls,
+            'thumbnailUrl': image_urls[0] if image_urls else None
+        }
+
+class ProductImage(db.Model):
+    __tablename__ = 'product_images'
+    id = db.Column(db.Integer, primary_key=True)
+    filename = db.Column(db.String(200), nullable=False)
+    product_id = db.Column(db.Integer, db.ForeignKey('products.id'), nullable=False)
+
+class User(db.Model):
+    __tablename__ = 'users'
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(100), nullable=False, unique=True)
+    email = db.Column(db.String(120), nullable=False, unique=True)
+    password_hash = db.Column(db.String(128), nullable=False)
+    is_admin = db.Column(db.Boolean, nullable=False, default=False)
+    phone_number = db.Column(db.String(50), nullable=True)
+
+    def to_dict(self):
+        return {
+            "id": self.id,
+            "username": self.username,
+            "email": self.email,
+            "is_admin": self.is_admin,
+            "phoneNumber": self.phone_number
+        }
+
+class Order(db.Model):
+    __tablename__ = 'orders'
+    id = db.Column(db.Integer, primary_key=True)
+    date = db.Column(db.DateTime, nullable=False, default=db.func.current_timestamp())
+    total = db.Column(db.Float, nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
+    address_id = db.Column(db.Integer, db.ForeignKey('addresses.id'), nullable=False)
+    address = db.relationship('Address', backref='order', uselist=False)
+    products = db.relationship('OrderProduct', backref='order', lazy=True)
+
+class Address(db.Model):
+    __tablename__ = 'addresses'
+    id = db.Column(db.Integer, primary_key=True)
+    full_name = db.Column(db.String(150), nullable=False)
+    street_address = db.Column(db.String(200), nullable=False)
+    apartment_suite = db.Column(db.String(100), nullable=True)
+    city = db.Column(db.String(100), nullable=False)
+    state_province = db.Column(db.String(100), nullable=True)
+    postal_code = db.Column(db.String(20), nullable=False)
+    country = db.Column(db.String(100), nullable=False)
+    phone_number = db.Column(db.String(50), nullable=True)
+
+class OrderProduct(db.Model):
+    __tablename__ = 'order_products'
+    id = db.Column(db.Integer, primary_key=True)
+    quantity = db.Column(db.Integer, nullable=False)
+    unit_price = db.Column(db.Float, nullable=False)
+    order_id = db.Column(db.Integer, db.ForeignKey('orders.id'), nullable=False)
+    product_id = db.Column(db.Integer, db.ForeignKey('products.id'), nullable=False)
+    product = db.relationship('Product', backref='orders', lazy=True)
