+++ frontend.log
@@ -1,0 +1,9 @@
+
+> frontend@0.0.0 dev
+> vite
+
+
+  VITE v7.1.4  ready in 392 ms
+
+  ➜  Local:   http://localhost:5173/
+  ➜  Network: use --host to expose
