+++ frontend/.env.example
@@ -1,0 +1,5 @@
+# URL base para el backend de la API
+VITE_API_BASE_URL=http://127.0.0.1:5000
+
+# Clave de API de Google Maps para la funcionalidad de Places Autocomplete
+VITE_GOOGLE_MAPS_API_KEY=TU_CLAVE_DE_API_AQUI
