+++ frontend/nginx.conf
@@ -1,0 +1,19 @@
+server {
+  listen 80;
+  server_name localhost;
+
+  # Root directory for static files
+  root /usr/share/nginx/html;
+  index index.html index.htm;
+
+  # Serve main React file for all non-file requests
+  location / {
+    try_files $uri /index.html;
+  }
+
+  # Optional: Add headers to prevent caching issues
+  location ~* \.(?:css|js)$ {
+    expires 1y;
+    add_header Cache-Control "public";
+  }
+}
